+++ _Typings/gameApi/types/itemComponents/light.d.ts
@@ -129,6 +129,6 @@ declare module "game:type" {
     }
 
     type LightMethod = {
-        [key: string | symbol]: never
+        [key: PropertyKey]: never
     }
 }