+++ _Typings/gameApi/types/itemComponents/roadGenerator.d.ts
@@ -123,6 +123,6 @@ declare module "game:type" {
     }
 
     type RoadGeneratorMethod = {
-        [key: string | symbol]: never
+        [key: PropertyKey]: never
     }
 }