+++ _Typings/gameApi/types/settingsComponents/terrain.d.ts
@@ -3,6 +3,6 @@
 
 declare module "game:type" {
     type Terrain = {
-        [key: string | symbol]: never
+        [key: PropertyKey]: never
     }
 }